Fix boids example
Mostly just copied the WGSL from the current wgpu repo.
